update architecture of file, update discription, added Local search algorithms
